requirements.txt
anyio==4.0.0
APScheduler==3.6.3
asgiref==3.7.2
cachetools==4.2.2
certifi==2023.7.22
crispy-bootstrap4==2023.1
dj-database-url==2.1.0
Django==4.2.5
django-crispy-forms==2.1
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
packaging==23.2
Pillow==10.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2023.3.post1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tornado==6.1
typing_extensions==4.8.0
tzlocal==5.0.1
whitenoise==6.6.0
